Ajuste en migraciones para que se puedan subir sin problema de la fk

+++ backend/database/migrations/2025_11_20_214050_empleado.php
@@ -20,10 +20,15 @@ return new class extends Migration
                 $table->string('alias')->nullable();
                 $table->enum('estado', ['activo', 'inactivo'])->default('activo');
                 $table->unsignedBigInteger('grupo_id')->nullable();
-                $table->foreign('grupo_id')->references('id')->on('grupos')->onUpdate('cascade')->onDelete('restrict');
                 $table->timestamps();
                 $table->softDeletes();
             });
+
+            if (Schema::hasTable('grupos')) {
+                Schema::table('empleados', function (Blueprint $table) {
+                    $table->foreign('grupo_id')->references('id')->on('grupos')->onUpdate('cascade')->onDelete('restrict');
+                });
+            }
     }
 
     /**
